Fix EIA 861 xlsx_maps to work back to 2001
Updated the paths in the file map to include the leading YYYY/ path
element that was missing in many of them prior to 2010, resulting in the
Excel files being unavailable during the extract step.

Fixed the skiprows for 2002 green_pricing_eia861 -- it was skipping the
header row, which meant the first row of data was being interpreted as
column names, which was breaking stuff, and losing that row of data.

Also updated the service_territory_eia861 transform function to account
for a couple of NA values which were getting cast to floats, preventing
addfips from parsing the columns correctly.