Normalize the Balancing Authority EIA 861 table
This splits the `balancing_authority_eia861` table into two tables, and
adds information from the data tables.

`balancing_authority_eia861` is an annual entity table which contains:
    * report_date (primary key)
    * balancing_authority_id_eia (primary key)
    * balancing_authority_code_eia (can be null)
    * balancing_authority_name_eia (can be null)

`balancing_authority_assn_eia861` is an association table
indicating which states and utilities have been seen in combination
which each other and with each balancing authority. These associations
are harvested from the data tables (`sales_eia861`,
`demand_response_eia861` etc.) which contain examples of these
associations. The table contains:
    * report_date (primary key)
    * balancing_authority_id_eia (primary key)
    * utility_id_eia (nullable primary key field)
    * state (nullable primary key field)

The association table depends on having an already (mostly) transformed
version of all the input tables available, so it has to be generated
after the main transform step. It also requires the denormalized
(original) version of the `balancing_authority_eia861` table, so it has
to be created in a step that follows the main transform loop.  Not sure
whether this should remain within the `eia861` module, or end up in the
harvesting phase later.