Consolidate Util/BA geometry code in new module
Commit adds a pudl.analysis.service_territory.py module, and migrates a
bunch of code related to generating and visualizing utility and
balancing authority territories, as inferred from the EIA 861 data, and
augmented with US Census DP1 county geometries. A few of these functions
were relocated from the `pudl.analysis.demand_mapping` module.

I also added a `tool.autopep8` section to the configuration
pyproject.toml file, to ensure that the local autopep8 and flake8
configurations are the same (allowing lines up to 88 characters long)
since a lot of people seem to like making the lines a little longer, and
I've come across a bunch of folks who recommend making this amendment to
the PEP 8 guidelines (and it's what the "black" code formatter has
adopted).