Simplify and generalize eia_code categorization
Simplified and generalized the `categorize_eia_code()` function to allow either
balancing authorities or utilities to take precedence, and to allow a specific
set of utility and balancing authority IDs to be passed in. This will let us
compile a set of individual utility IDs that we want to override the default
balancing authority IDs if need be -- using all BA and all Util IDs and setting
the priority to Utility just means everything (except 4 entities) are
interpreted as utilities which is... clearly not right.

Closes #628

Other minor changes:

* Changed the loglevels for log output in the Excel extractor to use a variety
  of debug, info, warning, so that we can have more or less output as desired
  -- rather than having basically everything at info, per feedback from
  @ptvrigo.
* Updated `pudl-dev` environment to use `sphinx>=3.0` since it no longer causes
  problems with our docs build, and it's been out of many months now.
* Updated pre-commit hooks to use isort 5.1.0
* Added a jupyterlab build script under devtools to install extensions & build
  the application with a single command. Useful for getting all the extensions
  back after recreating your conda environment. We should talk about how to
  make this not user-specific.