Add datastore to PUDL output object for EIA 861
With the new datastore setup, we need to have a datastore instance which
is part of the pudl output object, in order to support the interim ETL
for EIA 861 (and soon, FERC 714) since they are currently pulling the
raw data and transforming it, rather than accessing the PUDL DB like the
other output functions.

Also fixed a FERC 714 transform bug that seems to have arisen from
changes to how pandas deals with timezone aware datetime fields.