Simplify datapkg_to_sqlite script
Merging datapackages is very fast, and this script is really just about loading
a bundle of datapackages into SQLite (hence the name) so I've simplified the
options to focus on doing just those things, and to only store the merged
data package temporarily, in an automatically generated temporary directory,
until the SQLite load is finished.

This change was inspired by accidentally wiping out an entire pudl-work
directory including all downloaded data by giving the script a bad argument
for the output directory path in conjunction with the --clobber arg!

Now --clobber only pertains to the sqlite database file.

Closes #711