Use LocalResult to persist/cache results.
WIP: Naming of these artifacts still feels somewhat clunky and there
doesn't seem to be a good way to control the output path for this.

It is also unclear how to swap LocalResult for cloud-based storage, but
perhaps this is also not entirely necessary until after we have
kubernetes based deployment (if this is needed/useful).