Eliminate data chokepoint from the pipeline.
merge_dataframe_map task is convenient but dangerous when used on a
massive dataset. This is the case for epacems and caused dask workers
responsible for aggregating the data to crash repeatedly and clogged
my machine.

This refactoring moves the csv writing into the pipeline code such that
it is done on the mapped data and only table names are returned back
to the metadata builder step.

Also added --dask-executor-address that allows us to use pre-existing
dask cluster to run the stuff on and modified the code so that the right
executor is passed to the etl (rather than passing boolean flags). This
should allow more flexibility in picking the executor.