Ensure deterministic checksums on csv.gz outputs.
gzip files are adding mtimes in the headers which results in
non-deterministic checksums of the resulting files. This change
adds a workaround to ensure that mtime is not set and this should
allow us to generate compressed datapackages with deterministic
checksums.

Unfortunately, due to the existing known-bug, the workaround is somewhat
dirty and requires us to pass the iobuffer wrapped zip files that we
have to open by hand with the necessary parameters instead of relying
on the pandas/gzip integration.

See https://github.com/pandas-dev/pandas/issues/28103 for more details.