Integrate new Datastore with ETL.
This refactors the ETL part of the code to use the new Datastore API.

Dataset specific implementations of datastore are now implemented as
wrappers that provide extra functionality. This makes the passing of
datastore somewhat easier because we can create one instance of generic
datastore that can be passed to DatasetPipeline objects that may need
it and they can optionally wrap it.

epaipm datastore logic has been consolidated under extract/epaipm.py and
some constants have been moved into this module.