Get API_KEY_EIA when used rather than as constant
Previously the API_KEY_EIA value was being set as a global variable at
the top of the module. This meant that the value of the constant
reflected the value at time of module import. So if the user imports
pudl (e.g.  in a notebook) and then **later** sets the environment
variable, the value of the constant doesn't reflect that setting,
meaning the API key is usually None.

Obtaining the API key from `os.environ.get('API_KEY_EIA') when the key
is used inside the functions avoids this issue.