Investigate why EPA CEMS is sooo dang slooow
After seeing the speedups that @ezwelty was able to eek out of the FERC
714 ETL process and missing value imputations, I've become suspicious
that the EPA CEMS ETL does not need to take anywhere near as long as it
currently does.  I did some poking around to see where the bottleneck
is, and it seems to be the datastore -- not the extract or transform.
While I haven't tracked down the issue, I did move some of the constants
that pertain only to the EPA CEMS ETL into that module, and out of the
constants.py graveyard.