Modify ETL to use DataFrameCollection.
Replace use of dictionaries containing named pd.DataFrame instance
with DataFrameCollection. This class behaves similarly but stores
the underlying data frames on disk either locally or on GCS.

This reduces memory footprint of the ETL when run on single machine
(and probably also distributed, because prefect may use in-memory
caching even then) and also allows the pipeline to be run in a
distributed fashion where task will be exchanging references to
data frames that live in a shared cloud storage (GCS bucket).