Bugfixes for states=[ALL] and SQLite DB clobber check
Attempts to address issue #870 which notes we're checking for the
existence of an SQLAlchemy database URL, not a file path, when looking
to see if we're clobbering an existing PUDL database.

Also fixes a bug in which the special case of doing "ALL" states in the
EPA CEMS processing was failing, since they were listed as dictionary
keys, instead of a list or tuple.