Update fast_out fixture to use datastore fixture
The interim ETL functions for eia861 and ferc714 that are (temporarily,
hackishly) embedded within the PUDL output object need to know where to
look for the input data, which is a different place in CI than
(typically) on our local machines. This commit tells that the fast_out
object which is used for output testing (including running the interim
ETL functions for eia861 and ferc714) to look at whatever datastore is
within the pudl_datastore_fixture. Which still might not work because of
the hardcoded PUDL_YML in pudl.workspace.datastore but we'll see.