Add default Datastore to PudlTabl output objects
For incompletely integrated data sources like ferc714 and eia861, we
need to have access to a datastore within the PUDL output objects, so we
can read the raw input files and run the interim ETL process. With this
commit you can either pass in a Datastore object to use, or you leave it
set to None, the __init__ method will use the default Datastore
indicated by the user's .pudl.yml PUDL_IN path.