Variety of code improvements to the prefect ETL.
1. Added --pipeline-cache-path that controls where DataFrameCollections
and prefect task results should be stored.
2. Added target specifications to most of the tasks to allow
checkpointing
3. Refactored metadata construction such that writing of csv files,
computation of checksum and building of resource json descriptor are
done in parallel. This extracts get_tabular_data_resource out of
generate_metadata.
4. If datapackage contains epacems resources, skip goodtables
validations because it would choke the system to death.
5. Excel extractor is now native prefect task that pulls
datastore from context and can be run separately.
6. Simplified glue pipeline code (moved @task annotation into
ferc1_eia.glue)

This seems to finally be working fine.