gcs datastore cache will always be read-only
In order to separate responsibilities and avoid accidental
corruption, ETL will always use GCS cache for datastore as
read-only and will never attempt to modify it.

This will be enforced by permissions when run on cloud but
it is a good idea to always do this.

If we need to modify the content of the cache, we should run
pudl_datastore CLI to update it.