Write epacems parquet files directly to gcs.
When --upload-to-gcs is set, don't even bother writing to
local disk.

Ultimately, the etl should be able to set datapkg_dir and parquet_dir
to either local destination or to gcs path and everything should work
the same regardless of where the files are being written to, but we
are not there yet.

See https://github.com/catalyst-cooperative/pudl/issues/886 for more
context.