Simplify dependencies and update dev setup docs.
Rather than specifying all of the PUDL dependencies both in environment.yml and
in the setup.py, I'm now installing the PUDL package for development directly
within the environment.yml file (no more pip install -e ./). The only packages
that are still being specified independently in the conda environment are ones
that depend on precompiled binaries that aren't immediately available via the
pip install, or packages that we use for development but that aren't really part
of the pudl-dev environment. There are also some project specific packages
like recordlinkage.

Simplified and updated the development environment setup documentation.