Simplify use of live databases in testing
Remove the --live_pudl_db flag and replace with --live-dbs which will use
existing user PUDL and FERC1 SQLite databases to run the tests, rather than
generating new, temporary databases for testing purposes. This is primarily
intended for use with the data validation tests at the moment, but it can
also be used with e.g. the output tests or any other tests which depend on
having a database present. Note that some of the database dependent tests
can take quite a while to run against a full multi-year database, instead of
a single-year database intended for testing purposes.