Create exhaustive integration test settings
In order to be able to reliably run some multi-year tests (e.g. checking that
the database schemas we are generating from all of the different years of
FERC Form 1 database are mutually compatible) we do sometimes need to run the
tests against the *full* set of all years and tables of data. So I've added
a settings file specifying the full complement of data, which can be used
via pytest with the --etl-settings flag, or by doing `tox -e full_integration`
To just check the FERC 1 db schemas, `tox -e ferc1_schema` -- this does require
you to have all the FERC 1 raw input data (or it'll get downloaded for you)

Because we can tell the tests to run whatever ETL we want with --etl-settings
I removed the standalone test_ferc1_solo_etl test, and instead specified the
data that it should attempt to load in a ferc1-solo-test.yml file under
test/settings. This is run prior to the main integration tests by Tox.

Added a couple of basic tests to the ferc1_etl and pudl_engine tests, which
had just been `pass` statements. Now they at least check to see that the
fixture is of type `sa.engine.Engine` and check that a couple of tables which
should always be present appear in the engine.table_names() list.

Made some adjustments to which files are being included in coverage to more
accurately reflect how well we're doing.

Made the `ci` testenv the default -- if you just run `tox` that's what will
get run.  You can run `tox -av` to see the list of all the available test
environments with short descriptions of what they do.

Still need to finish documenting these changes via #940

Closes #942