Implement conversion of Census DP1 data to SQLite
Rather than the ad-hoc ETL that we were using to extract the zip archive
of the Census DP1 GeoDataBase into a local folder, this commit treats the
DP1 as a standard data source, and converts it into an SQLite DB alongside
the PUDL and FERC 1 DBs in the PUDL_OUT/sqlite directory. It uses GDAL's
`ogr2ogr` command line utility to do this, calling it as a subprocess. This
means that this utility will need to be installed on your system for it to
work. Our `pudl-dev` conda environment will install it for you, or it can
be installed in /usr/bin/ogr2ogr with `apt install gdal-bin`. Other non-conda
installation options are certainly possible but... it'll get complicated to
cover all of them. Hopefully this setup will work fine in our Docker and CI
environments, and anyone else who is using it on a desktop will use the
conda environment.

I did not end up using the SpatiaLite extensions to SQLite, since they weren't
necessary to get GeoPandas reading out of the DB just fine, and they introduce
some complexity into how one reads data out of the DB with SQLAlchemy/pandas.

The baby DP1 ETL is housed in a module / script called
`src/pudl/convert/censusdp1tract_to_sqlite.py` since we have a couple of other
`_to_sqlite` scripts in there, so it seemed conceptually simple. It's just a
single function, wrapping `ogr2ogr` so that it knows where to get the inputs
from (using the Datastore) and send the outputs to (the sqlite dir) based on
the `pudl_settings` dictionary.

Also added installation of `gdal-bin` to the setup for our CI. We'll see if it
actually works!

Closes #920