Bump python-snappy to back to 0.6.0
One source of weirdness in our conda environment seems to have been the
lack of python-snappy 0.6.0 on conda-forge (but availability on the
default channel). I poked the developers / packagers and got them to
update the conda-forge packaging for 0.6.0, so now it's available, and
the pudl-dev environment works with both conda and mamba.