Add historical instructions for EIA & FERC Forms
The instructions and blank forms associated with the collection of the
data in PUDL are often very helpful for interpreting the sometimes messy
contents of the database and raw inputs. In order to better understand
what respondents were responding to historically, I've collected old
versions of some of these forms from the Internet Archive's Wayback
Machine. This isn't exhaustive, but it's a start.