Update Excel Extract test to reflect corrected file map
The Excel Extractor test was trying to lookup the file associated with
`plant_frame` in 2009, and that tab doesn't actually exist in that year
-- a fact that's just now reflected in the XLSX maps... so the test was
failing. I changed the page that it was trying to look up, and also
changed the pages several other tests were looking up, to be pages that
we actually depend on, rather than ones we've never looked at.