Integrate PUDL Unit ID assignment into generator output
Ad-hoc integration of more expansive PUDL Unit ID assignments into the
output functions for the EIA 860 generators. Eventually this
functionality should go into the boiler generator association process,
but it's going to require additional debugging and development, and we
don't want to have to run the ETL over and over again to test it out.

This is still a work in progress. Outstanding questions / tasks...
* Move explanatory text from the prototype notebook into the docstrings.
* Consider lumping vs. splitting in the steam turbine fuel-based unit
  assignments, as well as the "simple" per-generator unit ID
  assignments.
* There's some code duplication between the different kinds of unit ID
  assignments. Is it worth trying to simplify?
* Should this process be optional in the gens_eia860() output function?
  It takes a minute or two to run. Alternatively, can it be made much
  faster?
* What kinds of tests or diagnostics can we use to understand whether
  these particular unit ID assignments are a good idea or not?