Update available data & ETL to include EPA CEMS 2020
* Update the DOI for EPA CEMS to point at new Zenodo archive including 2020 data.
* Update data_years and working_parts to include 2020 data for EPA CEMS
* Update fuel heat content validation bounds to reflect observed ranges.
* Note that the expected number of rows has not been updated in the data validations.
* Add 2020 to epacems_years in the etl_full.yml settings file

Surprisingly, there were no new missing plants in the 2020 EPA CEMS data which needed
to be added to the lat/lon cheat sheet. However when we update the EIA 860/923 to
include the 2020 data and start using that as the standalone test year, we'll probably
need to add some new entries, since the older years of EIA data won't be included in the
tests.

Data validations were run and updated to reflect the now available data, with the
exception of the expected number of records in each table, and the known outstanding
errors (an all NA column in the MCOE output)

Closes #1073