Add PUDL Unit IDs to test and fix dtype assignment errors
I added a gens_eia860() test to the fast output tests, and have it pull
the data with unit_ids=True to test the Unit ID assignment code.

I also merged in the dev branch, which included a change in the
generators_eia860() function, to use the column dtype standardization
function from pudl.helpers. But... this caused errors in the ID
assignment process -- a few hundred records get lost along the way if
all those dtypes are set. So... I reverted to my simpler dtype
assignment code. But this is confusing.