Move dtype assignment to end of generators_eia860()
The PUDL Unit ID assignment proces has unexpected behavior if it's operating on a
dataframe with the canonical datatypes (as set by convert_cols_dtypes()). The only
columns that are involved in the process are:

* plant_id_eia
* generator_id
* report_date
* unit_id_pudl
* bga_source
* fuel_type_code_pudl
* prime_mover_code

The unexpected behavior is that a few hundred rows are just... lost in the process. This
is caught by the check which ensures the index of the input and outputdataframes are
identical in the unit ID assignment function.

To work around this, I moved the convert_cols_dtypes() call to the very end of the
generators_eia860() function, after the unit ID assignments have taken place. But it
would be good to understand exactly why this is happening. Seems weird.