Update available data years and instructional PDFs
I updated the available years for datasets that have expanded since this
PR was made, and also added links to the other blank forms /
instructions that we've archived for EIA & FERC data as references.

I also standardized the names of the RST files that are in the data
dictionaries directory, since they were both different and redundant, if
those docs are within a dedicated data dictionaries directory.