Add output date frequency tests; parametrize output tests
Added a parametrized test that checks the date frequency of output
dataframes, relative to the gens_eia860 dataframe, to ensure that we
have annual / monthly dataframes as expected. This test is currently
expected to fail because of the bug referenced in issue #1088.

Also consolidated some of the "fast output tests" using parametrization,
and improved coverage to include all of the potential output ferc1,
eia860, eia923, and mcoe output tables.