WIP: Prep inputs for fuel allocation and allocate gens that aren't in the bf table.
Mostly amended the iterim outputs for the existing net generation 
allocation methodology to be compatible for the fuel allocation. This 
involved including the boiler fuel table and the necessary groupby/sums 
for the boiler for allocation into associate_generator_tables() and 
prep_alloction_fraction().

There are two simple tests in here for these "not in BF" generators. 
More need to be added.

I also need to implement a solution for the generators that do show up 
in the boiler fuel table. Easiest option would be to not allocate to the 
generators - rather aggregate to the units which the data is effectively 
already being reported in. The downside to this option is that it would 
make the process less compatible with the net generation allocation - 
the outputs are all in generator-based tables.