Replace janky merge_on_date_year with built-in merge_asof()
Tracked down the source of the frequency mismatch between some MCOE output dataframes
and the frequency of their output objects. It was due to some bad merges in both the EIA
923 outputs and the mcoe analysis outputs. Several problems arose from the fact that
our merge_on_date_year() function would silently choose just the left dataframe's
version of any column which appeared in both dataframes, leading to unexpected outcomes
when we didn't realize that the column existed in both the left and right dataframes.

I got rid of merge_on_date_year() entirely, as well as the is_annual() function which
was only called by it, and replaced them with a much shorter, simpler function in
helpers called clean_merge_asof() that uses the pandas.merge_asof() method to bring
together dataframes with different time frequencies. It also ensures that the columns
which are being merged on are sorted, as required by merge_asof(), and that they have
the same datatypes (e.g. if plant_id_eia is an int64 in one dataframe, and
pandas.Int64Dtype() in the other, it'll cast them both to a uniform type, which must be
explicitly specified as an input). It also ensures that report_date is a Datetime and
not just a column of iso date strings like "2021-07-27"

In addition to the frequency mismatch issue, replacing merge_on_date_year() surfaced a
bad bug in the fuel_receipts_costs_eia923() output function, which was merging in the
coalmines_eia923 table, including a column named just "state" -- i.e. the state that the
coalmine is in... not the state that the plant is in. Then when the plant's state was
merged in, it was ignored, since a state column already existed in the (at that point)
left dataframe. This probably had two effects:

* Many fewer records in the frc_eia923 table had states associated with them than would
  be expected, since not all fuel deliveries come from mines.
* The records that *did* end up having states had the states of the *mine* not the
  plant. This bad state field was then used to merge in the monthly average fuel prices
  from the EIA API in cases where no fuel price was reported.
* As a result, we were unable to fill in as many missing fuel prices as with might
  otherwise have been able to.

In addition, I discovered that the pudl_out.bga() and pudl_out.bga_eia860() methods,
which were linked up to the pudl.outputs.mcoe.boiler_generator_assn() and
pudl.outputs.glue.boiler_generator_assn_eia860() functions respectively, were identical,
except that the bga() version dropped the autoincrementing pseudo key field "id". I
think this situation is leftover from when we hadn't yet integrated our more complete
bga() process into the database, and it was being run on the fly by the output object.
So far as I know, this arrangement wasn't causing any actual errors, but it was
confusing and duplicative, and so I removed the bga() functionality, since bga_eia860()
follows the naming convention of other database tables more closely.

I moved the integration of the unit_id_pudl field into the boiler_fuel_eia923() and
generation_eia923() output functions, since those fields end up getting used later on,
and the were being added in several different places. This led to column conflicts and
duplicated code.

Added a Sphinx extension that will allow us to add re-directs in the documentation
if/when we move a page, and used it to redirect anyone visiting the new removed
data_dictionary.html page to the new page under data_dictionaries/pudl_db.html

Closes #1088
Closes #1102