Fix many-to-many merge in pudl.analysis.mcoe.heat_rate_by_gen()
There was one monthly-to-annual (or annual-to-annual, depending on the frequency of the
pudl_out aggregation) merge required a many-to-many merge, in the calculation of the
heat rates by generator. On one side, we had heat rates by unit potentially at monthly
resolution. And on the other side we had generator IDs per unit, but only at annual
resolution. The goal was to get heat rates by generator at monthly resolution,
broadcasting the per-unit heat rates across all generators within each unit for each
month. This isn't something that merge_asof() can do.

Instead, I created categorical year and month fields, and did a normal many-to-many
merge, broadcasting months in one direction, and generator IDs in the other.  It seems
like there *must* be a clean way to do this using time series and a multi index with
forward filling or something. Or maybe period indices but... after a couple of hours of
trying to figure that out I gave up. This *seems* to do the right thing, and should work
for monthly and annual cases at least, which is all we have now.

Other smaller changes:
* Applied convert_cols_dtypes() to the outputs of each MCOE function for uniformity,
  since it doesn't seem to break anything and avoids needing to set the data types
  inside the functions independently on some occasions.
* Enforce that the PudlTabl class can only have freq in (None, "MS", "AS")
* Add some type checking for other arguments which initialize the PudlTabl class (the
  pudl_engine and Datastore)
* Improved error message and logging output in pudl.validate.check_(min,max)_rows
  functions.
* Update the expected number of records in the EIA 923 output tables. Most of them only
  changed by a fraction of a percent.
* Update the expected number of records in the MCOE output tables. These were off by
  The number of monthly records for hr_by_gen and fuel_cost both increased by 3.7%,
  which is large compared to all the other changes. Is this unexpected?
* Added a CITATION.cff file to the repository, to provide easy citation access on
  GitHub, through new functionality they've added.