Address null columns in mcoe data validation
Created a list of deprecated columns which are skipped in the mcoe null
column validation, since they only exist in 2006 and earlier data, and
we can only calculate the MCOE for 2008 and later.

Updated the null column validation to check all columns before raising
an exception, so you can get a list of all of them in the error message.

Updated release notes to remove this piece of errata.

Closes #943