Use canonical PUDL data types to type columns & dataframes
Rather than passing data types in explicitly to
pudl.helpers.clean_merge_asof(), pass in which data source the columns
are from (e.g. "eia" or "ferc1") and lookup the correct data type in
pudl.constants.column_dtypes.

To do this concisely and readably, I made a couple of little helper
functions that look up the type of a single column, or a dictionary of
{column: data source} values. I inserted these into a variety of other
places where we were hard-coding data types for use with astype().

Also made the docstring for clean_merge_asof() more concise.