make generation allocation output mirror the standard generation table.
The cleanup of the clean_merge_as_of() function for the outputs was not 
working when the allocate_net_gen flag is True. The gens table had 
different columns than the allocated gen. The previous merges in the 
mcoe function were more "grab all of the columns every every single 
time".