Dependency and data release script updates
There's a bug in JupyterLab 3.1.6-3.1.7 that causes notebook cells to
get skipped, hidden, or deleted when running all cells, or individual
cells. This was happening in the pudl-examples notebooks when I was
trying to compile the data release for v0.4.0. I pinned the environment
in that repo to 3.1.0-3.1.4, and have done so in the environment file
here as well.

Also checking in the databeta.sh script as it was used to generate the
data release which is currently uploading to Zenodo.