Explicitly set boiler_generator_assn_eia860 dtypes
Something changed in pandas v1.3 such that the data types in the boiler
generator association table weren't matching the expected types anymore,
in particular plant_id_eia was showing up as a float rather than an int.
This was able to happen because in the main top level ETL function, we
set the types of the data tables, and then do the entity resolution /
harvesting, and then set the data types of the entity tables, but we
never re-set the types of the BGA table, which is actually built in the
midst EIA transform process.

However, the error that was actually showing up was in the
goodtables_pandas tabular data package validation, because the way that
errors are reported changed at some point, and so our attempt to parse
and create a short summary of the errors was failing. I hacked something
workable but temporary there, since we're going to remove the data
packages altogether shortly.

Closes #1070