Update data validation tests and revert datapackage validation
Update expected number of rows in the data validation tests for the BGA and BGA
dependent MCOE outputs to reflect the altered BGA table, which has been changed to have
more consistent data types throughout the process, and to work with pandas 1.3.x  The
BGA table has 4 more rows in it now. Not a big change.

Revert to not specifying a number of rows to validate with goodtables_pandas. The
library automatically checks all rows itself, but runs with limit_rows=1 to test headers
and other non-data aspects of the data resources. See the discussion in this issue:

https://github.com/ezwelty/goodtables-pandas-py/issues/4

Closes #1070