Improve error reporting on extra / missing Excel columns
Previously we checked whether the **number** of columns extracted from
Excel spreadsheets was what we expected, and if not, reported which
extra columns were found in the data, but not in our maps of the data.
But this is only one possible way that things can go wrong.

I've updated the check to compare the actual columns of the extracted
dataframe with those that we expect to see as sets, so if they're
different but have the same number of elements, we'll still be notified.

I've also updated it to report what partition of the data the error is
happening in, and to report both when there are **extra** columns in the
data that don't show up in our maps, and when there are columns
**missing** from the data that we expect to see, based on our maps. And
it's possible to get both kinds of warnings at the same time, since
we're looking at set differences, rather than just the number of
elements.