Add a "nuke" test environment to tox to rebuild and test everything
Added a "tox -e nuke" option which will:
* Run the CI tests as normal, and if they pass...
* Run the full ETL clobbering your ferc1 & pudl SQLite DBs and EPA CEMS
  Parquet datasets, using the `etl_full.yml` settings file.
* Run all of the database dependent tests all over again, but on the
  newly created live databases.
* Run the data validation tests.

This is all the stuff that needs to get run before merging `dev` into
`main` and provides a one-liner for rebuilding everything from scratch
and testing it completely. In total if everything runs successfully it
should take about 3.5 hours to run.

It will also generate a coverage report so you can see just how much of
the codebase is really being tested.