Move respondent_frequency to plants_eia860 (annual) table
So far as I can tell, whether a plant has to respond annually ore
monthly depends on capacity of the plant, and we know that can change
over time, so I've put the `respondent_frequency` field into the annual
plants entity table (`plants_eia860`).

Note that some of these changes will need to be reconciled with the
`dedupe-metadata` branch when it gets merged in, since where the
metadata is being stored has changed over there.

It looks like we're now running into the fact that there are plant names
/ utilities that haven't been mapped for 2020, as the FERC 1 tables are
failing with foreign key constraint errors.