EIA & FERC1 ETL works with all years through 2020
The full ETL with all FERC1 and EIA 860/923 data will run without
obvious errors. There are still tests and validations that fail, but at
least you can load the DB.

This does *not* include eia860m or EPA CEMS data yet. FERC-714 and
EIA-861 also remain to be updated for 2020.

Issues that remain:
* Something screwy is going on with FERC respondent 542 -- it shows up
  only in the `f1_respondent_id` table, and has all Null data there...
  and our unmapped utility finder script failed to identify it.  See
  #1304
* A defensive assertion aimed at identifying human errors in the ID
  mapping sheet is failing because (probably?) we have a fair number of
  plants and utilities with IDs but no names in there now. See #1305 and
  also #1232