Update eia860m, ferc714, eia923, eia861 for 2020
I ran the full ETL with eia860m integrated to get the next round of IDs for mapping, and
found that there was only 1 additional unmapped EIA utility and no additional plants. I
also created a new eia860m archive with the newest data (through 2021-08). It seems to
"just work"

I also updated the eia861 Zenodo archive to include the 2020 data, and switched the DOIs
over to the 2020-inclusive archives for ferc714 and eia861.

For eia861 I've updated the filemap.csv for 2020, but the column maps and other
per-table metadata files still need to be updated.

For ferc714 I updated the filenames and it gets through "extract" and raises and
AssertionError in the transform step for the hourly demand by planning area, which will
require some actual debugging.

I also created a new archive for the eia923, which saw updates on October 8th.
Thankfully it didn't break anything big -- just changed file named and added the
balancing_authority_code_eia to gf, bf, frc, and gen tables (it gets harvested out).