Update to flake8 v4.0; always install pudl for Tox
Remove the pin to flake8<4.0.0 since all of the plugins have now been
updated to work with the new release.

To avoid this kind of issue in the future, instead of separately
specifying dependencies within tox.ini, *always* install the PUDL
package for every specified testenv, so that whatever versions of
dependencies are specified in setup.py are used.