Simplify generator technology description filling routine
Pulled the technology description filling routine out of the generators_eia860
output function to stand alone, and simplified the backfilling portion of it.

Made filling tech description the default in the PudlTabl objects and generators_eia860
output function, as it's a straightforward fix that should be generally useful. This
also means the code will get tested now.

I checked that the output of this version and the previously checked in versions were
identical. It's just faster and easier to read.

There are still some cases where we could be filling in additional records.  For example
looking at plant_id_eia==470 (Comanche) units 1 & 2, they switch from SUB to BIT as
their primary fuel in like 2005, and this method, which requires consistent fuel types,
doesn't catch that. I wonder if we might be able to do something with prime movers, or
grouping of the energy source codes such that all of the coals are treated the same?